pyproject.toml
[project]
name = "projet-phil"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "openai>=2.15.0",
    "openpyxl>=3.1.5",
    "pymupdf>=1.26.7",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "streamlit>=1.53.0",
]
